В вычислении факториала изменил тип i на double
Тип int мог вызвать ошибку переполнения

+++ src/main/java/mitasov/calc/CompiledExpression.java
@@ -23,7 +23,7 @@ class CompiledExpression {
 
         double ret = 1;
 
-        for (int i = 1; i <= value; ++i) {
+        for (double i = 1; i <= value; ++i) {
             ret *= i;
         }
         return ret;
